Code cleanup; Added release version to login screen
