+++ webserver/core/hardware_layers/opi_zero2.cpp
@@ -73,25 +73,20 @@ void initializeHardware()
     //set pins as input
     for (int i = 0; i < MAX_INPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_inputs, ARRAY_SIZE(ignored_bool_inputs), i))
-        {
-            pinMode(inBufferPinMask[i], INPUT);
-        }
+        pinMode(inBufferPinMask[i], INPUT);
     }
 
     //set pins as output
     for (int i = 0; i < MAX_OUTPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-            pinMode(outBufferPinMask[i], OUTPUT);
+        pinMode(outBufferPinMask[i], OUTPUT);
     }
     // pwmSetRange(1024);
     // pwmSetClock(1);
     // //set PWM pins as output
     // for (int i = 0; i < MAX_ANALOG_OUT; i++)
     // {
-    // 	if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), analogOutBufferPinMask[i])) {
-    //     	pinMode(analogOutBufferPinMask[i], PWM_OUTPUT);
+    //      pinMode(analogOutBufferPinMask[i], PWM_OUTPUT);
     // 	}
     // }
 
@@ -117,8 +112,7 @@ void updateBuffersIn()
     //INPUT
     for (int i = 0; i < MAX_INPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_inputs, ARRAY_SIZE(ignored_bool_inputs), i))
-            if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = digitalRead(inBufferPinMask[i]);
+        if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = digitalRead(inBufferPinMask[i]);
     }
 
     pthread_mutex_unlock(&bufferLock); //unlock mutex
@@ -136,14 +130,12 @@ void updateBuffersOut()
     //OUTPUT
     for (int i = 0; i < MAX_OUTPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-            if (bool_output[i/8][i%8] != NULL) digitalWrite(outBufferPinMask[i], *bool_output[i/8][i%8]);
+        if (bool_output[i/8][i%8] != NULL) digitalWrite(outBufferPinMask[i], *bool_output[i/8][i%8]);
     }
     //ANALOG OUT (PWM)
     // for (int i = 0; i < MAX_ANALOG_OUT; i++)
     // {
-    //     if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), i))
-    // 		if (int_output[i] != NULL) pwmWrite(analogOutBufferPinMask[i], (*int_output[i] / 64));
+    //     if (int_output[i] != NULL) pwmWrite(analogOutBufferPinMask[i], (*int_output[i] / 64));
     // }
 
     pthread_mutex_unlock(&bufferLock); //unlock mutex
