+++ webserver/core/hardware_layers/pixtend2s.cpp
@@ -564,15 +564,13 @@ void updateBuffersIn()
     //DIGITAL INPUT
     for (int i = 0; i < MAX_DIG_IN; i++)
     {
-        if (pinNotPresent(ignored_bool_inputs, ARRAY_SIZE(ignored_bool_inputs), i))
-            if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = bitRead(InputData.byDigitalIn, i);
+        if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = bitRead(InputData.byDigitalIn, i);
     }
     
     //GPIO INPUT
     for (int i = MAX_DIG_IN; i < MAX_DIG_IN+MAX_GPIO_IN; i++)
     {
-        if (pinNotPresent(ignored_bool_inputs, ARRAY_SIZE(ignored_bool_inputs), i))
-            if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = bitRead(InputData.byGPIOIn, i-MAX_DIG_IN);
+        if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = bitRead(InputData.byGPIOIn, i-MAX_DIG_IN);
     }
     
     // uint8_t byFirmware;
@@ -593,49 +591,44 @@ void updateBuffersIn()
     {
         if (i < MAX_ANALOG_IN)
         {
-            if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                if (int_input[i] != NULL) *int_input[i] = analogInputs[i];
+            if (int_input[i] != NULL) *int_input[i] = analogInputs[i];
         }
         if ((i >= MAX_ANALOG_IN) && ( i < MAX_ANALOG_IN+MAX_TEMP_IN)) 
         {
-            if (i == MAX_ANALOG_IN){
-                if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                    if (int_input[i] != NULL) *int_input[i] = InputData.wTemp0;
+            if (i == MAX_ANALOG_IN)
+            {
+                if (int_input[i] != NULL) *int_input[i] = InputData.wTemp0;
             }
-            if (i == (MAX_ANALOG_IN+1)){
-                if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                    if (int_input[i] != NULL) *int_input[i] = InputData.wTemp1;
+            if (i == (MAX_ANALOG_IN+1))
+            {
+                if (int_input[i] != NULL) *int_input[i] = InputData.wTemp1;
             }
-            if (i == (MAX_ANALOG_IN+2)){
-                if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                    if (int_input[i] != NULL) *int_input[i] = InputData.wTemp2;
+            if (i == (MAX_ANALOG_IN+2))
+            {
+                if (int_input[i] != NULL) *int_input[i] = InputData.wTemp2;
             }
-            if (i == (MAX_ANALOG_IN+3)){
-                if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                    if (int_input[i] != NULL) *int_input[i] = InputData.wTemp3;
+            if (i == (MAX_ANALOG_IN+3))
+            {
+                if (int_input[i] != NULL) *int_input[i] = InputData.wTemp3;
             }
         }
         if ((i >= (MAX_ANALOG_IN+MAX_TEMP_IN)) && ( i < (MAX_ANALOG_IN+MAX_TEMP_IN+MAX_HUMID_IN))) 
         {
             if (i == (MAX_ANALOG_IN+MAX_TEMP_IN))
             {
-                if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                    if (int_input[i] != NULL) *int_input[i] = InputData.wHumid0;
+                if (int_input[i] != NULL) *int_input[i] = InputData.wHumid0;
             }
             if (i == ((MAX_ANALOG_IN+MAX_TEMP_IN)+1))
             {
-                if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                    if (int_input[i] != NULL) *int_input[i] = InputData.wHumid1;
+                if (int_input[i] != NULL) *int_input[i] = InputData.wHumid1;
             }
             if (i == ((MAX_ANALOG_IN+MAX_TEMP_IN)+2))
             {
-                if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                    if (int_input[i] != NULL) *int_input[i] = InputData.wHumid2;
+                if (int_input[i] != NULL) *int_input[i] = InputData.wHumid2;
             }
             if (i == ((MAX_ANALOG_IN+MAX_TEMP_IN)+3))
             {
-                if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-                    if (int_input[i] != NULL) *int_input[i] = InputData.wHumid3;
+                if (int_input[i] != NULL) *int_input[i] = InputData.wHumid3;
             }
         }
             
@@ -662,8 +655,7 @@ void updateBuffersOut()
     {
         if (i < MAX_DIG_OUT)
         {
-            if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-                if (bool_output[i/8][i%8] != NULL) bitWrite(OutputData.byDigitalOut, i, *bool_output[i/8][i%8]);
+            if (bool_output[i/8][i%8] != NULL) bitWrite(OutputData.byDigitalOut, i, *bool_output[i/8][i%8]);
         }  
     }
     
@@ -672,8 +664,7 @@ void updateBuffersOut()
     {
         if ((i >= MAX_DIG_OUT) && (i < (MAX_DIG_OUT+MAX_REL_OUT)))
         {
-            if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-                if (bool_output[i/8][i%8] != NULL) bitWrite(OutputData.byRelayOut, i-MAX_DIG_OUT, *bool_output[i/8][i%8]);
+            if (bool_output[i/8][i%8] != NULL) bitWrite(OutputData.byRelayOut, i-MAX_DIG_OUT, *bool_output[i/8][i%8]);
         }
     }
     
@@ -682,8 +673,7 @@ void updateBuffersOut()
     {
         if ((i >= MAX_DIG_OUT+MAX_REL_OUT) && (i < (MAX_DIG_OUT+MAX_REL_OUT+MAX_GPIO_OUT)))
         {
-            if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-                if (bool_output[i/8][i%8] != NULL) bitWrite(OutputData.byGPIOOut, i-(MAX_DIG_OUT+MAX_REL_OUT), *bool_output[i/8][i%8]);
+            if (bool_output[i/8][i%8] != NULL) bitWrite(OutputData.byGPIOOut, i-(MAX_DIG_OUT+MAX_REL_OUT), *bool_output[i/8][i%8]);
         }
     }
 
@@ -696,18 +686,15 @@ void updateBuffersOut()
     {
         if (i < 2)
         {
-            if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), i))
-                if (int_output[i] != NULL) analogOutputs[i] = (*int_output[i] / 64);
+            if (int_output[i] != NULL) analogOutputs[i] = (*int_output[i] / 64);
         }
         else
         {
-            if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), i))
-                if (int_output[i] != NULL) pwmOutputs[i-2] = *int_output[i];
+            if (int_output[i] != NULL) pwmOutputs[i-2] = *int_output[i];
         }
     }
     // PWM0Ctrl1L - PWM0Ctrl1H
-    if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), 4))
-        if (int_output[4] != NULL) OutputData.wPWM0Ctrl1 = *int_output[4];
+    if (int_output[4] != NULL) OutputData.wPWM0Ctrl1 = *int_output[4];
     
     // UCCtrl0
     if (byte_output[0] != NULL) OutputData.byUCCtrl0 = *byte_output[0];
