+++ webserver/core/hardware_layers/raspberrypi_old.cpp
@@ -74,28 +74,23 @@ void initializeHardware()
     //set pins as input
     for (int i = 0; i < MAX_INPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_inputs, ARRAY_SIZE(ignored_bool_inputs), i))
+        pinMode(inBufferPinMask[i], INPUT);
+        if (i != 0 && i != 1) //pull down can't be enabled on the first two pins
         {
-            pinMode(inBufferPinMask[i], INPUT);
-            if (i != 0 && i != 1) //pull down can't be enabled on the first two pins
-            {
-                pullUpDnControl(inBufferPinMask[i], PUD_DOWN); //pull down enabled
-            }
+            pullUpDnControl(inBufferPinMask[i], PUD_DOWN); //pull down enabled
         }
     }
 
     //set pins as output
     for (int i = 0; i < MAX_OUTPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-            pinMode(outBufferPinMask[i], OUTPUT);
+        pinMode(outBufferPinMask[i], OUTPUT);
     }
 
     //set PWM pins as output
     for (int i = 0; i < MAX_ANALOG_OUT; i++)
     {
-        if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), i))
-            pinMode(analogOutBufferPinMask[i], PWM_OUTPUT);
+        pinMode(analogOutBufferPinMask[i], PWM_OUTPUT);
     }
 }
 
@@ -119,8 +114,7 @@ void updateBuffersIn()
     //INPUT
     for (int i = 0; i < MAX_INPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_inputs, ARRAY_SIZE(ignored_bool_inputs), i))
-            if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = digitalRead(inBufferPinMask[i]);
+        if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = digitalRead(inBufferPinMask[i]);
     }
 
     pthread_mutex_unlock(&bufferLock); //unlock mutex
@@ -138,15 +132,13 @@ void updateBuffersOut()
     //OUTPUT
     for (int i = 0; i < MAX_OUTPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-            if (bool_output[i/8][i%8] != NULL) digitalWrite(outBufferPinMask[i], *bool_output[i/8][i%8]);
+        if (bool_output[i/8][i%8] != NULL) digitalWrite(outBufferPinMask[i], *bool_output[i/8][i%8]);
     }
 
     //ANALOG OUT (PWM)
     for (int i = 0; i < MAX_ANALOG_OUT; i++)
     {
-        if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), i))
-            if (int_output[i] != NULL) pwmWrite(analogOutBufferPinMask[i], (*int_output[i] / 64));
+        if (int_output[i] != NULL) pwmWrite(analogOutBufferPinMask[i], (*int_output[i] / 64));
     }
 
     pthread_mutex_unlock(&bufferLock); //unlock mutex
