+++ webserver/core/hardware_layers/unipi.cpp
@@ -172,21 +172,16 @@ void initializeHardware()
 
     for (int i = 0; i < MAX_OUTPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-            pinMode(DOUT_PINBASE + i, OUTPUT);
+        pinMode(DOUT_PINBASE + i, OUTPUT);
     }
 
     for (int i = 0; i < MAX_INPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_inputs, ARRAY_SIZE(ignored_bool_inputs), i))
-        {
-            pinMode(inputPinMask[i], INPUT);
-            pullUpDnControl(inputPinMask[i], PUD_UP);
-        }
+        pinMode(inputPinMask[i], INPUT);
+        pullUpDnControl(inputPinMask[i], PUD_UP);
     }
 
-    if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), ANALOG_OUT_PIN))
-        pinMode(ANALOG_OUT_PIN, PWM_OUTPUT);
+    pinMode(ANALOG_OUT_PIN, PWM_OUTPUT);
 
     pthread_t ADCthread;
     pthread_create(&ADCthread, NULL, readAdcThread, NULL);
@@ -211,15 +206,13 @@ void updateBuffersIn()
     pthread_mutex_lock(&bufferLock); //lock mutex
     for (int i = 0; i < MAX_INPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_inputs, ARRAY_SIZE(ignored_bool_inputs), i))
-            if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = !digitalRead(inputPinMask[i]); //printf("[IO%d]: %d | ", i, !digitalRead(inputPinMask[i]));
+        if (bool_input[i/8][i%8] != NULL) *bool_input[i/8][i%8] = !digitalRead(inputPinMask[i]); //printf("[IO%d]: %d | ", i, !digitalRead(inputPinMask[i]));
     }
 
     //printf("\nAnalog Inputs:");
     for (int i = 0; i < 2; i++)
     {
-        if (pinNotPresent(ignored_int_inputs, ARRAY_SIZE(ignored_int_inputs), i))
-            if (int_input[i] != NULL) *int_input[i] = mcp_adcRead(i); //printf("[AI%d]: %d | ", i, mcp_adcRead(i));
+        if (int_input[i] != NULL) *int_input[i] = mcp_adcRead(i); //printf("[AI%d]: %d | ", i, mcp_adcRead(i));
     }
     //printf("\n");
 
@@ -238,12 +231,10 @@ void updateBuffersOut()
     //printf("\nDigital Outputs:\n");
     for (int i = 0; i < MAX_OUTPUT; i++)
     {
-        if (pinNotPresent(ignored_bool_outputs, ARRAY_SIZE(ignored_bool_outputs), i))
-            if (bool_output[i/8][i%8] != NULL) digitalWrite(DOUT_PINBASE + i, *bool_output[i/8][i%8]); //printf("[IO%d]: %d | ", i, digitalRead(DOUT_PINBASE + i));
+        if (bool_output[i/8][i%8] != NULL) digitalWrite(DOUT_PINBASE + i, *bool_output[i/8][i%8]); //printf("[IO%d]: %d | ", i, digitalRead(DOUT_PINBASE + i));
     }
     
-    if (pinNotPresent(ignored_int_outputs, ARRAY_SIZE(ignored_int_outputs), 0))
-        if(int_output[0] != NULL) pwmWrite(ANALOG_OUT_PIN, (*int_output[0] / 64));
+    if(int_output[0] != NULL) pwmWrite(ANALOG_OUT_PIN, (*int_output[0] / 64));
     
     pthread_mutex_unlock(&bufferLock); //unlock mutex
 }
